Check-in form update, autofill personal data, booked room selected default, autofill special requests, payment summary added

+++ desktop/desktop-app/Models/BillingItem.cs
@@ -8,12 +8,12 @@ namespace Hotel_erp_Winforms_App.Models
     {
         public DateTime Date { get; set; }
         public string Description { get; set; }
-        public int UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public int Quantity { get; set; }
         public decimal Tax { get; set; }
         public decimal Total { get; set; }
 
-        public BillingItem(DateTime date, string description, int unitPrice, int quantity, decimal tax, decimal total)
+        public BillingItem(DateTime date, string description, decimal unitPrice, int quantity, decimal tax, decimal total)
         {
             Date = date;
             Description = description;
