+++ desktop/desktop-app/Models/Service.cs
@@ -9,7 +9,7 @@ namespace Hotel_erp_Winforms_App.Models
 
     public class Service
     {
-        public Service(int id, string nameHu, string desciptionHu, ServiceTypeHu selectedServiceTypeHu, int price, string nameEn, string desciptionEn, ServiceTypeEn selectedServiceTypeEn)
+        public Service(int id, string nameHu, string desciptionHu, ServiceTypeHu selectedServiceTypeHu, decimal price, string nameEn, string desciptionEn, ServiceTypeEn selectedServiceTypeEn)
         {
             Id = id;
             NameHu = nameHu;
@@ -25,7 +25,7 @@ namespace Hotel_erp_Winforms_App.Models
         public string NameHu { get; set; }
         public string DesciptionHu { get; set; }
         public ServiceTypeHu SelectedServiceTypeHu { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public string NameEn { get; set; }
         public string DesciptionEn { get; set; }
         public ServiceTypeEn SelectedServiceTypeEn { get; set; }
